added better support for tooltips in radio/select dialogs, and added tooltips to DNA export dialog explaining the options
